Updated GenericDialog
updated the GenericDialog component to have a new event based system. This works by the calling component sending an event title as a parameter. Once a value has been chosen by the the dialog, the appropriate event will be issued along the same event bus that triggered the dialog, for the caller to respond to.